requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
kafka-python>=2.0.2
websockets>=10.0
python-multipart>=0.0.5